Do not remote WiresharkOuiService contribute to liveness probe checks
It requires calling an external service,
which might not be available sometimes,
causing the pod to be considered not live and restarted by K8s too many times.
We are instead relying on the default Quarkus liveness check, returning UP when the application is started.
We still have a readiness probe, that will rely on data returned by remote OUI Services.